cmd/requirements.txt
requests
bs4
instaloader
rich
distro
flask
qrcode
pillow
